规划路径和实时路径三维显示

+++ src/main/java/com/org/iuabc/entity/PathCoordinate.java
@@ -22,11 +22,11 @@ public class PathCoordinate {
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     private Long pointId;
 
-    private Float pointX;
+    private Double pointX;
 
-    private Float pointY;
+    private Double pointY;
 
-    private Float pointZ;
+    private Double pointZ;
 
     private Long pathId;
